refactor: 코드컨벤션 수정 및 안 쓰이는 필드 제거
