+++ backend/selection/src/main/java/com/selection/repository/ArticleRepository.java
@@ -4,5 +4,4 @@ import com.selection.domain.article.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ArticleRepository extends JpaRepository<Article, Long> {
-
 }
