+++ backend/selection/src/main/java/com/selection/repository/NoticeReposiotry.java
@@ -4,5 +4,4 @@ import com.selection.domain.notice.Notice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NoticeReposiotry extends JpaRepository<Notice, Long> {
-
 }
