+++ backend/selection/src/main/java/com/selection/repository/QuestionRepository.java
@@ -4,5 +4,4 @@ import com.selection.domain.question.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface QuestionRepository extends JpaRepository<Question, Long> {
-
 }
