+++ backend/selection/src/main/java/com/selection/repository/TagRepository.java
@@ -4,5 +4,4 @@ import com.selection.domain.tag.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface TagRepository extends JpaRepository<Tag, Long> {
-
 }
