+++ backend/selection/src/main/java/com/selection/repository/UserRepository.java
@@ -11,5 +11,4 @@ public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
 
     Boolean existsByEmail(String email);
-
 }
