+++ backend/selection/src/main/java/com/selection/service/article/ArticleService.java
@@ -5,8 +5,6 @@ import com.selection.domain.article.Article;
 import com.selection.dto.article.ArticleResponse;
 import com.selection.dto.article.ArticleSaveRequest;
 import com.selection.repository.ArticleRepository;
-import com.selection.repository.QuestionRepository;
-import com.selection.repository.TagRepository;
 import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -16,9 +14,6 @@ import org.springframework.stereotype.Service;
 public class ArticleService {
 
     private final ArticleRepository articleRepository;
-    private final TagRepository tagRepository;
-    private final QuestionRepository questionRepository;
-
 
     @Transactional
     public ArticleResponse create(ArticleSaveRequest requestDto) {
